Fixes and tweaks

+++ Api/Domain/SearchResult.cs
@@ -2,4 +2,4 @@ namespace SearchHitCount.Domain;
 
 public record SearchResult(
     string Input,
-    int Hits);
+    long Hits);
